contrib/acl: fix unmatched init checks for ACL database
This resolves an issue where the setup timestamp did not match the
one that was attempting to be retrieved at init time. While some
databases did not have an issue with this, e.g. the local database,
others that prevent duplicate k/v insertion such as postgres do.

Relevant discussion:
https://lakefs.slack.com/archives/C016726JLJW/p1728288654114759